+++ src/tests/mocks/data.ts
@@ -2,6 +2,7 @@ import mongoose from 'mongoose';
 
 const userOneId = new mongoose.Types.ObjectId();
 const userTwoId = new mongoose.Types.ObjectId();
+const userThreeId = new mongoose.Types.ObjectId();
 
 const tagOneId = new mongoose.Types.ObjectId();
 const tagTwoId = new mongoose.Types.ObjectId();
@@ -14,6 +15,9 @@ const tutorialThreeId = new mongoose.Types.ObjectId();
 const trackOneId = new mongoose.Types.ObjectId();
 const trackTwoId = new mongoose.Types.ObjectId();
 const trackThreeId = new mongoose.Types.ObjectId();
+
+const feedbackOneId = new mongoose.Types.ObjectId();
+const feedbackTwoId = new mongoose.Types.ObjectId();
 
 // User Data
 export const users = [
@@ -34,6 +38,15 @@ export const users = [
 		googleID: '55667788',
 		isAdmin: false,
 		isSuperAdmin: false
+	},
+	{
+		_id: userThreeId,
+		email: 'user3@example.com',
+		name: 'User3',
+		displayPicture: '',
+		googleID: '12121212',
+		isAdmin: true,
+		isSuperAdmin: true
 	}
 ];
 
@@ -172,5 +185,29 @@ export const tracks = [
 			userId: users[1]._id
 		},
 		isApproved: true
+	}
+];
+
+// Feedback Data
+export const feedbacks = [
+	{
+		_id: feedbackOneId,
+		title: 'Feedback One Title',
+		message: 'Feedback One Message',
+		isRead: false,
+		submittedBy: {
+			name: users[1].name,
+			userId: users[1]._id
+		}
+	},
+	{
+		_id: feedbackTwoId,
+		title: 'Feedback Two Title',
+		message: 'Feedback Two Message',
+		isRead: true,
+		submittedBy: {
+			name: users[1].name,
+			userId: users[1]._id
+		}
 	}
 ];