+++ src/tests/mocks/seed.ts
@@ -2,13 +2,12 @@ import User from '../../models/User';
 import Tag from '../../models/Tag';
 import Tutorial from '../../models/Tutorial';
 import Track from '../../models/Track';
+import Feedback from '../../models/Feedback';
 
-import { users, tags, tutorials, tracks } from './data';
+import { users, tags, tutorials, tracks, feedbacks } from './data';
 
 export const populateUsers = (done: jest.DoneCallback) => {
-	const userOne = new User(users[0]).save();
-	const userTwo = new User(users[1]).save();
-	Promise.all([userOne, userTwo]).then(() => done());
+	User.create(users).then(() => done());
 };
 
 export const removeUsers = (done: jest.DoneCallback) => {
@@ -37,4 +36,12 @@ export const populateTracks = (done: jest.DoneCallback) => {
 
 export const removeTracks = (done: jest.DoneCallback) => {
 	Track.deleteMany({}).then(() => done());
+};
+
+export const populateFeedbacks = (done: jest.DoneCallback) => {
+	Feedback.create(feedbacks).then(() => done());
+};
+
+export const removeFeedbacks = (done: jest.DoneCallback) => {
+	Feedback.deleteMany({}).then(() => done());
 };